:style: Atualiza o estilo do arquivo hb.c

+++ hb.c
@@ -61,7 +61,7 @@ hbtransform(XftGlyphFontSpec *specs, const Glyph *glyphs, size_t len, int x, int
 	int start = 0, length = 1, gstart = 0;
 	hb_codepoint_t *codepoints = calloc(len, sizeof(hb_codepoint_t));
 
-	for (int idx = 1, specidx = 1; idx < len; idx++) {
+	for (size_t idx = 1, specidx = 1; idx < len; idx++) {
 		if (glyphs[idx].mode & ATTR_WDUMMY) {
 			length += 1;
 			continue;
@@ -85,7 +85,7 @@ hbtransform(XftGlyphFontSpec *specs, const Glyph *glyphs, size_t len, int x, int
 	hbtransformsegment(specs[start].font, glyphs, codepoints, gstart, length);
 
 	/* Apply the transformation to glyph specs. */
-	for (int i = 0, specidx = 0; i < len; i++) {
+	for (size_t i = 0, specidx = 0; i < len; i++) {
 		if (glyphs[i].mode & ATTR_WDUMMY)
 			continue;
 		if (glyphs[i].mode & ATTR_BOXDRAW) {
